+++ docs/release-notes.md
@@ -3,6 +3,7 @@
 - Hover on folder button shows folder location
 - senpcli: Always show units/time instead of time/units in progress bars
 - senpcli: Add anime tracking functionality
+- senpcli: Add option to print direct download links instead of downloading
 
 # Bug fixes
 - Fix some gogo downloads randomly failing, if you still experience it be sure to report it